requirements.txt
requests
htmlparser
bs4
wget
ipwhois